Start of Client Side Controllers (#9)
* Add files via upload

A template with a bunch of backgrounds and a standardized top bar to build off of.

* Add files via upload

Additions:
+Fun Owlz Events and Venues Css and Html Totally done
+Added more templates
+Full Screen Popup Template and Popup Control Java-script

Comments:
Feel free to use any of the styles I have added to use on your pages.
If you look into venue-events.css there is more styles available.
Most of the generic styles are in templateStyles.css and the old styles are unchanged, so you should be safe to do a direct change over without any modification.

* Add files via upload

* Add files via upload

First Version of the Server
Its running on Spring Boot with Thymeleaf as it's html formatter.
Id recommend installing Spring Tool Suite into eclipse before you try to work with it.

* Add files via upload

Modified the fun owl pages to generate info with javascript files. The ones the generate are labeled as - Experimental.